+++ OmiconShop/OmiconShop.Domain/Entities/Product.cs
@@ -12,7 +12,7 @@ namespace OmiconShop.Domain.Entities
 
         public string Description { get; set; }
 
-        public int Price { get; set; }
+        public decimal Price { get; set; }
 
         public string Type { get; set; }
